fix curriculum credit point data type
git-svn-id: http://argo.unfuddle.com/svn/argo_ois/trunc/bonga-bonga@76 dada7bbb-e600-4cd3-a4f0-d41eb9236a84

+++ loaders.py
@@ -137,7 +137,7 @@ class CurriculumLoader(bulkloader.Loader):
              ('level_of_education', get_dictionary_key),
              ('form_of_training', get_dictionary_key),
              ('nominal_years', int),
-             ('nominal_credit_points', float),
+             ('nominal_credit_points', long),
              ('degree', get_dictionary_key),
             ])
